bot_cadastro_pordutos/requirements.txt
botcity-framework-web
botcity-framework-core
pandas